revert original train_segmentor and make new file for new implementation
