+++ src/otx/algorithms/segmentation/configs/ocr_lite_hrnet_18_mod2/model.py
@@ -49,4 +49,4 @@ model = dict(
 load_from = "https://storage.openvinotoolkit.org/repositories/openvino_training_extensions\
 /models/custom_semantic_segmentation/litehrnet18_imagenet1k_rsc.pth"
 
-fp16 = dict(loss_scale=512.0)
+# fp16 = dict(loss_scale=512.0)
