+++ src/otx/recipes/stages/segmentation/incremental.py
@@ -2,16 +2,16 @@ _base_ = ["./train.py", "../_base_/models/segmentors/segmentor.py"]
 
 optimizer = dict(_delete_=True, type="Adam", lr=1e-3, eps=1e-08, weight_decay=0.0)
 
-optimizer_config = dict(
-    _delete_=True,
-    grad_clip=dict(
-        # method='adaptive',
-        # clip=0.2,
-        # method='default',
-        max_norm=40,
-        norm_type=2,
-    ),
-)
+# optimizer_config = dict(
+#     _delete_=True,
+#     grad_clip=dict(
+#         # method='adaptive',
+#         # clip=0.2,
+#         # method='default',
+#         max_norm=40,
+#         norm_type=2,
+#     ),
+# )
 
 log_config = dict(
     interval=10,
